migrate worker lease timing tests to failpoints
Replace the flaky semantic worker renew, lease-loss, and panic timing checks with deterministic failpoint-driven coverage so worker lease transitions can be validated without race-prone sleeps.

Made-with: Cursor
