+++ pkg/server/semantic_worker_test.go
@@ -1103,303 +1103,6 @@ func TestSemanticWorkerProcessEmbedTaskWithoutEmbedderDoesNotPanic(t *testing.T)
 	}
 }
 
-func TestSemanticWorkerRenewsLeaseForLongImageTask(t *testing.T) {
-	extractor := &gatedServerImageExtractor{
-		text:    "renewed image text",
-		started: make(chan struct{}, 1),
-		release: make(chan struct{}),
-	}
-	var releaseOnce sync.Once
-	releaseExtractor := func() {
-		releaseOnce.Do(func() { close(extractor.release) })
-	}
-	t.Cleanup(releaseExtractor)
-	b := newTestBackendForSemanticWorkerWithOptions(t, backend.Options{
-		DatabaseAutoEmbedding: true,
-		AsyncImageExtract: backend.AsyncImageExtractOptions{
-			Enabled:   true,
-			Workers:   1,
-			QueueSize: 8,
-			Extractor: extractor,
-		},
-	})
-	fileID := insertServerImageFileForExtractTest(t, b, "/img/renew.png", "image/png", []byte("fake-png"))
-	payload, err := json.Marshal(semantic.ImgExtractTaskPayload{Path: "/img/renew.png", ContentType: "image/png"})
-	if err != nil {
-		t.Fatal(err)
-	}
-	now := time.Now().UTC()
-	availableAt := now.Add(-time.Second)
-	if _, err := b.Store().EnqueueSemanticTask(context.Background(), &semantic.Task{
-		TaskID:          "img-task-renew",
-		TaskType:        semantic.TaskTypeImgExtractText,
-		ResourceID:      fileID,
-		ResourceVersion: 1,
-		Status:          semantic.TaskQueued,
-		MaxAttempts:     3,
-		AvailableAt:     availableAt,
-		PayloadJSON:     payload,
-		CreatedAt:       now,
-		UpdatedAt:       now,
-	}); err != nil {
-		t.Fatal(err)
-	}
-
-	leaseDuration := 80 * time.Millisecond
-	claimed, found, err := b.Store().ClaimSemanticTask(context.Background(), time.Now().UTC(), leaseDuration, semantic.TaskTypeImgExtractText)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if !found {
-		t.Fatal("expected image task to be claimable")
-	}
-
-	m := newSemanticWorkerManager(b, nil, nil, nil, SemanticWorkerOptions{LeaseDuration: leaseDuration})
-	if m == nil {
-		t.Fatal("expected semantic worker manager")
-	}
-	done := make(chan struct{})
-	go func() {
-		defer close(done)
-		m.processTask(context.Background(), &semanticTarget{
-			tenantID: semanticLocalTenantID,
-			backend:  b,
-			store:    b.Store(),
-		}, claimed)
-	}()
-
-	select {
-	case <-extractor.started:
-	case <-time.After(2 * time.Second):
-		t.Fatal("image task did not start")
-	}
-
-	renewedLeaseUntil := waitForNamedTaskLeaseAfter(t, b, claimed.TaskID, claimed.LeaseUntil, time.Second)
-	// Use a point just before the renewed lease to avoid precision/race flakes
-	// around the old lease boundary.
-	recovered, err := b.Store().RecoverExpiredSemanticTasks(context.Background(), renewedLeaseUntil.Add(-time.Millisecond), 64)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if recovered != 0 {
-		t.Fatalf("recovered=%d, want 0", recovered)
-	}
-	if renewedLeaseUntil.Before(claimed.LeaseUntil.UTC()) {
-		t.Fatalf("renewed lease_until=%v, want after %v", renewedLeaseUntil, claimed.LeaseUntil.UTC())
-	}
-
-	releaseExtractor()
-	select {
-	case <-done:
-	case <-time.After(2 * time.Second):
-		t.Fatal("processTask did not finish")
-	}
-
-	waitForNamedTaskStatus(t, b, claimed.TaskID, string(semantic.TaskSucceeded), 2*time.Second)
-	task := mustGetServerSemanticTask(t, b, claimed.TaskID)
-	if task.AttemptCount != 1 {
-		t.Fatalf("attempt_count=%d, want 1", task.AttemptCount)
-	}
-}
-
-func TestSemanticWorkerCancelsLeaseLostImageTaskWithoutAckOrRetry(t *testing.T) {
-	core, recorded := observer.New(zap.InfoLevel)
-	restoreLogger := logger.L()
-	logger.Set(zap.New(core))
-	t.Cleanup(func() { logger.Set(restoreLogger) })
-
-	extractor := &gatedServerImageExtractor{
-		text:     "stale image text",
-		started:  make(chan struct{}, 1),
-		release:  make(chan struct{}),
-		canceled: make(chan struct{}, 1),
-	}
-	var releaseOnce sync.Once
-	releaseExtractor := func() {
-		releaseOnce.Do(func() { close(extractor.release) })
-	}
-	t.Cleanup(releaseExtractor)
-	b := newTestBackendForSemanticWorkerWithOptions(t, backend.Options{
-		DatabaseAutoEmbedding: true,
-		AsyncImageExtract: backend.AsyncImageExtractOptions{
-			Enabled:   true,
-			Workers:   1,
-			QueueSize: 8,
-			Extractor: extractor,
-		},
-	})
-	fileID := insertServerImageFileForExtractTest(t, b, "/img/lease-lost.png", "image/png", []byte("fake-png"))
-	payload, err := json.Marshal(semantic.ImgExtractTaskPayload{Path: "/img/lease-lost.png", ContentType: "image/png"})
-	if err != nil {
-		t.Fatal(err)
-	}
-	now := time.Now().UTC()
-	if _, err := b.Store().EnqueueSemanticTask(context.Background(), &semantic.Task{
-		TaskID:          "img-task-lease-lost",
-		TaskType:        semantic.TaskTypeImgExtractText,
-		ResourceID:      fileID,
-		ResourceVersion: 1,
-		Status:          semantic.TaskQueued,
-		MaxAttempts:     3,
-		AvailableAt:     now.Add(-time.Second),
-		PayloadJSON:     payload,
-		CreatedAt:       now,
-		UpdatedAt:       now,
-	}); err != nil {
-		t.Fatal(err)
-	}
-
-	leaseDuration := 80 * time.Millisecond
-	claimed, found, err := b.Store().ClaimSemanticTask(context.Background(), time.Now().UTC(), leaseDuration, semantic.TaskTypeImgExtractText)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if !found {
-		t.Fatal("expected image task to be claimable")
-	}
-
-	m := newSemanticWorkerManager(b, nil, nil, nil, SemanticWorkerOptions{LeaseDuration: leaseDuration})
-	if m == nil {
-		t.Fatal("expected semantic worker manager")
-	}
-	done := make(chan struct{})
-	go func() {
-		defer close(done)
-		m.processTask(context.Background(), &semanticTarget{
-			tenantID: semanticLocalTenantID,
-			backend:  b,
-			store:    b.Store(),
-		}, claimed)
-	}()
-
-	select {
-	case <-extractor.started:
-	case <-time.After(2 * time.Second):
-		t.Fatal("image task did not start")
-	}
-
-	if _, err := b.Store().DB().Exec(`UPDATE semantic_tasks
-		SET status = ?, receipt = NULL, leased_at = NULL, lease_until = NULL, available_at = ?, updated_at = ?
-		WHERE task_id = ?`,
-		semantic.TaskQueued, now.Add(time.Second), now.Add(time.Second), claimed.TaskID); err != nil {
-		t.Fatal(err)
-	}
-
-	select {
-	case <-extractor.canceled:
-	case <-time.After(2 * time.Second):
-		t.Fatal("expected lease-lost cancellation")
-	}
-	select {
-	case <-done:
-	case <-time.After(2 * time.Second):
-		t.Fatal("processTask did not finish after lease loss")
-	}
-
-	task := mustGetServerSemanticTask(t, b, claimed.TaskID)
-	if task.Status != string(semantic.TaskQueued) {
-		t.Fatalf("task status=%q, want %q", task.Status, semantic.TaskQueued)
-	}
-	if task.AttemptCount != 1 {
-		t.Fatalf("attempt_count=%d, want 1", task.AttemptCount)
-	}
-
-	waitForObservedLog(t, recorded, "semantic_worker_lease_lost", 2*time.Second)
-	assertNoObservedLogForTask(t, recorded, claimed.TaskID, "semantic_worker_ack_ok", "semantic_worker_ack_failed", "semantic_worker_retry_scheduled", "semantic_worker_retry_failed", "semantic_worker_dead_lettered")
-}
-
-func TestSemanticWorkerPanicStopsLeaseRenewal(t *testing.T) {
-	extractor := &panicServerImageExtractor{started: make(chan struct{}, 1)}
-	b := newTestBackendForSemanticWorkerWithOptions(t, backend.Options{
-		DatabaseAutoEmbedding: true,
-		AsyncImageExtract: backend.AsyncImageExtractOptions{
-			Enabled:   true,
-			Workers:   1,
-			QueueSize: 8,
-			Extractor: extractor,
-		},
-	})
-	fileID := insertServerImageFileForExtractTest(t, b, "/img/panic.png", "image/png", []byte("fake-png"))
-	payload, err := json.Marshal(semantic.ImgExtractTaskPayload{Path: "/img/panic.png", ContentType: "image/png"})
-	if err != nil {
-		t.Fatal(err)
-	}
-	now := time.Now().UTC()
-	if _, err := b.Store().EnqueueSemanticTask(context.Background(), &semantic.Task{
-		TaskID:          "img-task-panic",
-		TaskType:        semantic.TaskTypeImgExtractText,
-		ResourceID:      fileID,
-		ResourceVersion: 1,
-		Status:          semantic.TaskQueued,
-		MaxAttempts:     3,
-		AvailableAt:     now.Add(-time.Second),
-		PayloadJSON:     payload,
-		CreatedAt:       now,
-		UpdatedAt:       now,
-	}); err != nil {
-		t.Fatal(err)
-	}
-
-	leaseDuration := 80 * time.Millisecond
-	claimed, found, err := b.Store().ClaimSemanticTask(context.Background(), time.Now().UTC(), leaseDuration, semantic.TaskTypeImgExtractText)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if !found {
-		t.Fatal("expected image task to be claimable")
-	}
-
-	m := newSemanticWorkerManager(b, nil, nil, nil, SemanticWorkerOptions{LeaseDuration: leaseDuration})
-	if m == nil {
-		t.Fatal("expected semantic worker manager")
-	}
-	panicCh := make(chan any, 1)
-	done := make(chan struct{})
-	go func() {
-		defer close(done)
-		defer func() {
-			if r := recover(); r != nil {
-				panicCh <- r
-			}
-		}()
-		m.processTask(context.Background(), &semanticTarget{
-			tenantID: semanticLocalTenantID,
-			backend:  b,
-			store:    b.Store(),
-		}, claimed)
-	}()
-
-	select {
-	case <-extractor.started:
-	case <-time.After(2 * time.Second):
-		t.Fatal("image task did not start")
-	}
-	select {
-	case <-done:
-	case <-time.After(2 * time.Second):
-		t.Fatal("processTask did not panic and finish")
-	}
-	select {
-	case <-panicCh:
-	default:
-		t.Fatal("expected processTask panic to be recovered in test")
-	}
-
-	deadline := time.Now().Add(2 * time.Second)
-	for time.Now().Before(deadline) {
-		recovered, err := b.Store().RecoverExpiredSemanticTasks(context.Background(), time.Now().UTC(), 64)
-		if err != nil {
-			t.Fatal(err)
-		}
-		if recovered > 0 {
-			waitForNamedTaskStatus(t, b, claimed.TaskID, string(semantic.TaskQueued), time.Second)
-			return
-		}
-		time.Sleep(20 * time.Millisecond)
-	}
-	t.Fatal("expected expired task to be recoverable after handler panic")
-}
-
 func TestSemanticTaskLeaseExecutionStopIsIdempotent(t *testing.T) {
 	e := &semanticTaskLeaseExecution{
 		ctx:    context.Background(),
